Remove messageboxes and prompts from NotificationAdmin. Didn't remove the admin entirely because bridges use it as a pseudo-logger (probably want to switch to proper logging at some point). Fix a bunch more compiler warnings previously hidden by inclusion of folly headers.

+++ iModelCore/DgnPlatform/Tests/DgnProject/Published/CodeSpec_Test.cpp
@@ -15,13 +15,13 @@ USING_NAMESPACE_BENTLEY_DPTEST
 //========================================================================================
 struct CodeSpecTests : public DgnDbTestFixture
 {
-    void ValidateCodeFragmentSpec(CodeFragmentSpecCR, CodeFragmentSpec::Type, Utf8CP, bool, int minChars=0, int maxChars=CodeFragmentSpec::MAX_MaxChars);
+    void ValidateCodeFragmentSpec(CodeFragmentSpecCR, CodeFragmentSpec::Type, Utf8CP, bool, size_t minChars=0, size_t maxChars=CodeFragmentSpec::MAX_MaxChars);
 };
 
 //---------------------------------------------------------------------------------------
 // @bsimethod                                   Shaun.Sewall                    01/2017
 //---------------------------------------------------------------------------------------
-void CodeSpecTests::ValidateCodeFragmentSpec(CodeFragmentSpecCR fragmentSpec, CodeFragmentSpec::Type type, Utf8CP prompt, bool inSequenceMask, int minChars, int maxChars)
+void CodeSpecTests::ValidateCodeFragmentSpec(CodeFragmentSpecCR fragmentSpec, CodeFragmentSpec::Type type, Utf8CP prompt, bool inSequenceMask, size_t minChars, size_t maxChars)
     {
     BeAssert(fragmentSpec.IsValid());
     BeAssert(fragmentSpec.GetType() == type);
